requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiohttp==3.9.1
tenacity==8.2.3
python-gitlab==4.2.0
redis==5.0.1
APScheduler==3.10.4
cryptography==41.0.7
boto3==1.34.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
openai==1.6.1
